+++ src/tests/e2e/cases.rs
@@ -847,7 +847,10 @@ pub fn resize_terminal_window() {
             name: "wait for terminal to be resized and app to be re-rendered",
             instruction: |remote_terminal: RemoteTerminal| -> bool {
                 let mut step_is_complete = false;
-                if remote_terminal.cursor_position_is(53, 2) && remote_terminal.tip_appears() {
+                if remote_terminal.cursor_position_is(53, 2)
+                    && remote_terminal.tip_appears()
+                    && remote_terminal.snapshot_contains("Ctrl +")
+                {
                     // size has been changed
                     step_is_complete = true;
                 }